% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromPictoRdata.R
\name{FromPictoRdata}
\alias{FromPictoRdata}
\title{Transform a PNG into a matrix in R}
\usage{
FromPictoRdata(X, Y, save = FALSE)
}
\arguments{
\item{X}{the path to the file to analyses}

\item{Y}{the path to the folder containing named images of unique colors corresponding to a specific type of organism potentially present on the analysed image.}

\item{save}{\code{TRUE} to save an *.Rdata file containing the obtained matrix (default to \code{FALSE})}
}
\value{
A list containing a matrix with every cell value corresponding to a a specific type of organism present on the analyzed image and a table providing the legend.
}
\description{
This function allows you to transform an annotated PNG file into a matrix with each cell value corresponding to a a specific type of organism
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
