% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndPatchSize.R
\name{IndPatchSize}
\alias{IndPatchSize}
\title{Patch size distribution extractor}
\usage{
IndPatchSize(X, Y, Z = X)
}
\arguments{
\item{X}{the path of the file to analyses}

\item{Y}{the path of the folder containing the named pic which unique color fitting a specific type of organism potentially present on the plot}

\item{Z}{a chain of character to name the *.csv and *.jpg files (defaults to \code{X})}
}
\value{
A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image, also prints a picture with the labeled patches (same resolution as original)
}
\description{
This function allows you to extract the size and type of all the patches present in an image without using the GDAL library.
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
