% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndPatchSizePoly.R
\name{IndPatchSizePoly}
\alias{IndPatchSizePoly}
\title{Patch size distribution extractor}
\usage{
IndPatchSizePoly(X, Y, Z = X, pythonpath = NULL)
}
\arguments{
\item{X}{the path of the file to analyses}

\item{Y}{the path to the folder containing named images of unique colors corresponding to a specific type of organism potentially present on the analysed image.}

\item{Z}{a chain of character to name the *.csv and *.jpg files (defaults to \code{X})}
}
\value{
A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image, also prints a picture with the labeled patches (same resolution as original)
}
\description{
This function allows you to extract the size and type of all the patches present in an image using the GDAL library. WARNING: You need the Python library GDAL to use this function. Go to: \code{https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/} for more informations on how to proceed.
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
