% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchDistance.R
\name{PatchDistance}
\alias{PatchDistance}
\title{Distance bewteen patch}
\usage{
PatchDistance(X, Y = NA, Z = "your_mosaic", pathtopython = NULL,
  minsize = 0)
}
\arguments{
\item{X}{wether (i) the path of the file to analyse, (ii) a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}) or (iii) a SpatialPolygonDataFrame.}

\item{Y}{if \code{X} is the path of the file to analyse, \code{Y} is the path of the folder containing the named pic which unique color fitting a specific type of organism potentially present on the plot. If \code{Y=NA}, then \code{X} is considered being a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}).}

\item{Z}{a chain of character to name the *.csv and *.jpg files (defaults to \code{X})}
}
\value{
A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image, also prints a picture with the labeled patches (same resolution as original)
}
\description{
This function allows you to extract the distance bewteen patch as well as their sizes, excluding colonies which are too close from the edges (i.e. wether directly touching an edge or for which an edge is closer than the identified closest neighbor).
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
