% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyDistanceParr.R
\name{PolyDistanceParr}
\alias{PolyDistanceParr}
\title{Distance matrix bewteen polygons}
\usage{
PolyDistanceParr(X, cores = 1)
}
\arguments{
\item{X}{a list of polygon shape object.}

\item{cores}{the number of CPU cores to use for the analysis, defaults to 1 (CAUTION: each core will have a specific RAM requirement, take that in account very carefully).}
}
\value{
The distance matrix.
}
\description{
Generates the distance matrix between a set of polygons
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{distance}
\keyword{image}
\keyword{patch,}
\keyword{size,}
