% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pw_An_size_dis.R
\name{Pw_An_size_dis}
\alias{Pw_An_size_dis}
\title{Power Analysis for Size Distribution}
\usage{
Pw_An_size_dis(X, Y = NA, scale = 1, log10 = TRUE)
}
\arguments{
\item{X}{wether (i) the path of the file to analyse, (ii) a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}), (iii) a list containing the sizes of each organisms belonging to each group.}

\item{Y}{if \code{X} the path of the file to analyse, \code{Y} the path to the folder containing named images of unique colors corresponding to a specific type of organism potentially present on the analysed image. If \code{Y=NA}, then \code{X} is considered being a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}).}

\item{log10}{if \code{TRUE} the distributions are logged in base 10 (default to \code{TRUE})}

\item{Scale}{the size of the area covered by one pixel on the image (default to 1).}
}
\value{
A dataframe with the results of the power analysis for the 5 most commons metric used to describe a size distribution: mean, geometrical mean, kurtosis, skewness, coefficient of variation. The 2.5 and 97.5 percentile of the estimation are returned for each metric and for 10 increasing number of points.
}
\description{
This function allows you to proceed with a power analysis for the 5 most commons metrics used to describe a size distribution: mean, geometrical mean, kurtosis, skewness, coefficient of variation.
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{cover,}
\keyword{distributions.}
\keyword{image}
\keyword{patch}
\keyword{power}
\keyword{size}
\keyword{size,}
\keyword{species,}
