% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pw_An_VPC.R
\name{Pw_An_VPC}
\alias{Pw_An_VPC}
\title{Power Analysis for Point Counts}
\usage{
Pw_An_VPC(X, Y = NA, Cover = TRUE, Species = TRUE, Copower = 0.9,
  Cotol = 0.1, Sppower = 0.8, SpGoal = 0.5, speed = 100, max_n = 5000)
}
\arguments{
\item{X}{the path of the file to analyses}

\item{Y}{the path of the folder containing the named pic which unique color fitting a specific type of organism potentially present on the plot. If \code{Y=NA}, then \code{X} is considered being a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}).}

\item{Cover}{if \code{TRUE} then power analysis is runned for the cover estimation (default to \code{TRUE})}

\item{Species}{if \code{TRUE} then power analysis is runned for the number of species estimation (default to \code{TRUE})}

\item{Copower}{the desired power for the power analysis for the cover estimation (default to \code{0.9})}

\item{Cotol}{the tolerance for the cover estimation (default to \code{0.1})}

\item{Sppower}{the desired power for the power analysis for the number of species estimation (default to \code{0.8})}

\item{speed}{the speed of the estimation, faster is less precise, minimum is 1, default to 100.}

\item{max_n}{the maximum number of point allowed, default to 5000.}

\item{Spgoal}{the minimum ratio of the number of species to detect (default to \code{0.5})}
}
\value{
A dataframe with the results of the power analysis.
}
\description{
This function allows you to proceed with a power analysis of Point Counts on a annotated image.
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{cover,}
\keyword{image}
\keyword{patch}
\keyword{power}
\keyword{species,}
