% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sorting_raw.R
\name{Sorting_raw}
\alias{Sorting_raw}
\title{Sorting and renaming pictures}
\usage{
Sorting_raw(X, email = NA, erase = TRUE)
}
\arguments{
\item{X}{The path to the folder containing the pictures. Note: The name of this folder will be used to name the newly created fodlers and in the new pictures' name.}

\item{email}{the email to witch updates will be sent.}

\item{erase}{if \code{TRUE} the original folders cointaining the pictures are erased.}
}
\description{
This function renames and sorts pictures in different folders based on the focal length provided by the EXIF file.
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
