% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndPatchSize.R, R/IndPatchSize_old.R
\name{IndPatchSize}
\alias{IndPatchSize}
\title{Patch size distribution extractor}
\usage{
IndPatchSize(X, Y, Z)

IndPatchSize(X, Y, Z)
}
\arguments{
\item{X}{the path of the file to analyses}

\item{Y}{the path of the folder containing the named pic which unique color fitting a specific type of organism potentially present on the plot}

\item{Z}{a chain of character to name the *.csv and *.jpg files}

\item{X}{the path of the file to analyses}

\item{Y}{the path of the folder containing the named pic which unique color fitting a specific type of organism potentially present on the plot}

\item{Z}{a chain of character to name the *.csv and *.jpg files}
}
\value{
A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image, also prints a picture with the labeled patches (same resolution as original)

A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image, also prints a picture with the labeled patches (same resolution as original)
}
\description{
This function allows you to extract the size and type of all the patches present in an image

This function allows you to extract the size and type of all the patches present in an image
}
\examples{
#working on it
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
