% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatchDistance.R
\name{PatchDistance}
\alias{PatchDistance}
\title{Distance between patches}
\usage{
PatchDistance(X, Y = NA, Z = "your_mosaic", pathtopython = NULL,
  minsize = 0, lescoeur = 1, smoothing = 1)
}
\arguments{
\item{X}{wether (i) the path of the file to analyse, (ii) a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}) or (iii) a SpatialPolygonDataFrame.}

\item{Y}{if \code{X} the path of the file to analyse, \code{Y} the path to the folder containing named images of unique colors corresponding to a specific type of organism potentially present on the analysed image. If \code{Y=NA}, then \code{X} is considered being a matrix where each cell contains a number representing the cell type (as outputed by the function \code{FromPictoRdata}).}

\item{Z}{a chain of character to name the *.csv and *.jpg files (defaults to \code{X})}

\item{pathtopython}{specifies the path to the function \code{gdal_polygonize.py}}

\item{minsize}{the minimum size to consider a patch in the analysis (in number of pixels)}

\item{lescoeur}{the number of CPU cores to use for the analysis, defaults to 1 (CAUTION: each core will have a specific RAM requirement, take that in account very carefully).}

\item{smoothing}{a number between 1 and 100 defining the degree of polygonal smoothing to apply, with 1 meaning that 1% of the polygons node are kept, 100 meaning that 100% are kept. (a smaller number will compute faster and/or prevent bugs).}
}
\value{
A dataframe (both in R and as a *.csv in the working directory) containing the type and size of each patch present on the analyzed image and its distances to every other relevant patch.
}
\description{
This function allows you to extract the distance between patches as well as their sizes, excluding colonies which are too close from the edges (i.e. wether directly touching an edge or for which an edge is closer than the identified closest neighbor).
}
\examples{
#working on it
}
\keyword{Mosaic,}
\keyword{analysis,}
\keyword{image}
\keyword{patch,}
\keyword{size}
